% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BenMAP_population_shp.R
\name{read_BenMAP_population_shp}
\alias{read_BenMAP_population_shp}
\title{read_BenMAP_population_shp}
\usage{
read_BenMAP_population_shp(path, crs = NULL, verbose = getOption("verbose"))
}
\arguments{
\item{verbose}{}
}
\description{
read_BenMAP_population_shp
}
