% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_BenMAP_result_shp.R
\name{read_BenMAP_shp}
\alias{read_BenMAP_shp}
\title{read_BenMAP_result_shp}
\usage{
read_BenMAP_shp(
  path,
  crs = NULL,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{path}{(character) path to compressed ESRI shapefile}

\item{crs}{(optional) passed to [shptools::read_shp()]}

\item{verbose}{(logical)}
}
\description{
read_BenMAP_result_shp
}
